+++ packages/nodejs/src/storage/StorageBackendDiskAsync.ts
@@ -13,7 +13,7 @@ import {
     SupportedStorageTypes,
     toJson,
 } from "#general";
-import { mkdir, readdir, readFile, rm, writeFile } from "node:fs/promises";
+import { mkdir, open, readdir, readFile, rename, rm, writeFile } from "node:fs/promises";
 import { join } from "node:path";
 
 const logger = new Logger("StorageBackendDiskAsync");
@@ -58,6 +58,7 @@ export class StorageBackendDiskAsync extends MaybeAsyncStorage {
     async close() {
         this.isInitialized = false;
         await this.#finishAllWrites();
+        await this.#fsyncStorageDir();
     }
 
     filePath(fileName: string) {
@@ -66,6 +67,7 @@ export class StorageBackendDiskAsync extends MaybeAsyncStorage {
 
     async clear() {
         await this.#finishAllWrites();
+        await this.#fsyncStorageDir();
         await rm(this.#path, { recursive: true, force: true });
         await mkdir(this.#path, { recursive: true });
     }
@@ -134,12 +136,38 @@ export class StorageBackendDiskAsync extends MaybeAsyncStorage {
             return this.#writeFile(fileName, value);
         }
 
-        const promise = writeFile(this.filePath(fileName), value, "utf8").finally(() => {
+        // TODO write temp, fsync, move
+        const promise = this.#writeAndMoveFile(this.filePath(fileName), value).finally(() => {
             this.#writeFileBlocker.delete(fileName);
         });
         this.#writeFileBlocker.set(fileName, promise);
 
         return promise;
+    }
+
+    async #writeAndMoveFile(filepath: string, value: string): Promise<void> {
+        const tmpName = `${filepath}.tmp`;
+        await writeFile(tmpName, value, { encoding: "utf8", flush: true }); // flush does fsync after write
+        await rename(tmpName, filepath);
+    }
+
+    /**
+     * fsync on a directory ensures there are no rename operations etc. which haven't been persisted to disk.
+     * We do this as best effort to ensure that all writes are persisted to disk.
+     */
+    async #fsyncStorageDir() {
+        const fd = await open(this.#path, "r");
+        try {
+            await fd.sync();
+        } catch (error) {
+            // Windows will cause `EPERM: operation not permitted, fsync`
+            // for directories, so lets catch this generically
+            if ((error as any).code !== "EPERM") {
+                throw error;
+            }
+        } finally {
+            await fd.close();
+        }
     }
 
     async delete(contexts: string[], key: string) {